copied necessary fixes from trunk to v8 branch: * fixed crash on recurring appointments win null timezone (bug 7146) * OutlookTransportListener: fixed wrong method name (syncDataEnd -> sendDataEnd) * fixed winmaincpp.h and customization.h definitions
git-svn-id: https://outlook-client.forge.funambol.org/svn/outlook-client/branches/v8@169 1e7ea831-755a-0410-847c-f2bf8621b869

+++ mainclientDll/src/cpp/outlook/ClientContact.cpp
@@ -489,11 +489,11 @@ const wstring ClientContact::getComplexProperty(const wstring& propertyName) {
                 address2 = valueBuffer.substr(1, (valueBuffer.length()-2));
             }else{
                 address1 = valueBuffer.substr(0 , separator);
-                int separatorEnd = valueBuffer.find(extAdrSeparatorEnd);
+                unsigned long separatorEnd = valueBuffer.find(extAdrSeparatorEnd);
                 if (separatorEnd == StringBuffer::npos){
                     separatorEnd = valueBuffer.length();
                 }
-                if (separatorEnd < valueBuffer.length()){
+                if (separatorEnd < (int)(valueBuffer.length())){
                     address1.append(valueBuffer.substr(separatorEnd +1, (valueBuffer.length() - separatorEnd -1)));
                 }
                 address2 = valueBuffer.substr((separator + extAdrSeparatorLenght), (separatorEnd - separator - extAdrSeparatorLenght) );
